unittests adapted to new generic reconciler
