+++ pkg/3scale/amp/operator/apicast_reconciler_test.go
@@ -7,6 +7,7 @@ import (
 	"k8s.io/api/policy/v1beta1"
 
 	appsv1alpha1 "github.com/3scale/3scale-operator/pkg/apis/apps/v1alpha1"
+	"github.com/3scale/3scale-operator/pkg/reconcilers"
 	appsv1 "github.com/openshift/api/apps/v1"
 	imagev1 "github.com/openshift/api/image/v1"
 	routev1 "github.com/openshift/api/route/v1"
@@ -31,6 +32,9 @@ func TestApicastReconciler(t *testing.T) {
 		apicastManagementAPI       = "disabled"
 		oneValue             int64 = 1
 	)
+
+	ctx := context.TODO()
+
 	apimanager := &appsv1alpha1.APIManager{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      name,
@@ -79,11 +83,10 @@ func TestApicastReconciler(t *testing.T) {
 	cl := fake.NewFakeClient(objs...)
 	clientAPIReader := fake.NewFakeClient(objs...)
 
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	BaseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
+	baseReconciler := reconcilers.NewBaseReconciler(cl, s, clientAPIReader, ctx, log)
+	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseReconciler, apimanager)
 
-	apicastReconciler := NewApicastReconciler(BaseAPIManagerLogicReconciler)
+	apicastReconciler := NewApicastReconciler(baseAPIManagerLogicReconciler)
 	_, err = apicastReconciler.Reconcile()
 	if err != nil {
 		t.Fatal(err)
