--- pkg/3scale/amp/operator/deployment_config_reconciler_test.go
@@ -1,338 +1,0 @@
-package operator
-
-import (
-	"context"
-	"testing"
-
-	appsv1alpha1 "github.com/3scale/3scale-operator/pkg/apis/apps/v1alpha1"
-	"github.com/3scale/3scale-operator/pkg/helper"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	appsv1 "github.com/openshift/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-)
-
-func TestDeploymentConfigBaseReconcilerCreate(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	reconciler := NewDeploymentConfigBaseReconciler(baseAPIManagerLogicReconciler, NewCreateOnlyDCReconciler())
-
-	desired := &appsv1.DeploymentConfig{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "DeploymentConfig",
-			APIVersion: "apps.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myDC",
-			Namespace: namespace,
-		},
-	}
-
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{Name: "myDC", Namespace: namespace}
-	existing := &appsv1.DeploymentConfig{}
-	err = cl.Get(context.TODO(), namespacedName, existing)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestDeploymentConfigBaseReconcilerUpdateOwnerRef(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	existing := &appsv1.DeploymentConfig{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "DeploymentConfig",
-			APIVersion: "apps.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myDC",
-			Namespace: namespace,
-		},
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existing}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	reconciler := NewDeploymentConfigBaseReconciler(baseAPIManagerLogicReconciler, NewCreateOnlyDCReconciler())
-
-	desired := existing.DeepCopy()
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{Name: "myDC", Namespace: namespace}
-	reconciled := &appsv1.DeploymentConfig{}
-	err = cl.Get(context.TODO(), namespacedName, reconciled)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(reconciled.GetOwnerReferences()) != 1 {
-		t.Fatal("reconciled does not have owner reference")
-	}
-
-	if reconciled.GetOwnerReferences()[0].Name != name {
-		t.Fatalf("reconciled owner reference is not apimanager, expected: %s, got: %s", name, reconciled.GetOwnerReferences()[0].Name)
-	}
-}
-
-type myCustomDeploymentConfigReconciler struct {
-}
-
-func (r *myCustomDeploymentConfigReconciler) IsUpdateNeeded(desired, existing *appsv1.DeploymentConfig) bool {
-	existing.Spec.Replicas = 4
-	return true
-}
-
-func newmyCustomDeploymentConfigReconciler() DeploymentConfigReconciler {
-	return &myCustomDeploymentConfigReconciler{}
-}
-
-func TestDeploymentConfigReconcileUpdateNeeded(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	existing := &appsv1.DeploymentConfig{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "DeploymentConfig",
-			APIVersion: "apps.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myDC",
-			Namespace: namespace,
-		},
-	}
-	// existing does not need to be updated to set owner reference
-	err = controllerutil.SetControllerReference(apimanager, existing, s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existing}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	reconciler := NewDeploymentConfigBaseReconciler(baseAPIManagerLogicReconciler, newmyCustomDeploymentConfigReconciler())
-
-	desired := existing.DeepCopy()
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{Name: "myDC", Namespace: namespace}
-	reconciled := &appsv1.DeploymentConfig{}
-	err = cl.Get(context.TODO(), namespacedName, reconciled)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-	if reconciled.Spec.Replicas != 4 {
-		t.Fatalf("reconciled have reconciled data. Expected: 4, got: %d", reconciled.Spec.Replicas)
-	}
-}
-
-func TestDeploymentConfigReconcileReplicas(t *testing.T) {
-	logger := logf.Log.WithName("operator_test")
-	dcFactory := func() *appsv1.DeploymentConfig {
-		return &appsv1.DeploymentConfig{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "DeploymentConfig",
-				APIVersion: "apps.openshift.io/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "myDC",
-				Namespace: "myNS",
-			},
-			Spec: appsv1.DeploymentConfigSpec{
-				Replicas: 3,
-			},
-		}
-	}
-
-	cases := []struct {
-		testName       string
-		desired        func() *appsv1.DeploymentConfig
-		expectedResult bool
-	}{
-		{"NothingToReconcile", func() *appsv1.DeploymentConfig { return dcFactory() }, false},
-		{"ReplicasReconcile",
-			func() *appsv1.DeploymentConfig {
-				desired := dcFactory()
-				desired.Spec.Replicas = desired.Spec.Replicas + 1000
-				return desired
-			}, true,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.testName, func(subT *testing.T) {
-			existing := dcFactory()
-			update := DeploymentConfigReconcileReplicas(tc.desired(), existing, logger)
-			if update != tc.expectedResult {
-				subT.Fatalf("result failed, expected: %t, got: %t", tc.expectedResult, update)
-			}
-			if existing.Spec.Replicas != tc.desired().Spec.Replicas {
-				subT.Fatalf("replica reconciliation failed, existing: %d, desired: %d", existing.Spec.Replicas, tc.desired().Spec.Replicas)
-			}
-
-		})
-	}
-}
-
-func TestDeploymentConfigReconcileContainerResources(t *testing.T) {
-	logger := logf.Log.WithName("operator_test")
-	emptyResourceRequirements := corev1.ResourceRequirements{
-		Limits:   corev1.ResourceList{},
-		Requests: corev1.ResourceList{},
-	}
-	notEmptyResources := corev1.ResourceRequirements{
-		Limits: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse("100m"),
-			corev1.ResourceMemory: resource.MustParse("110Mi"),
-		},
-		Requests: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse("200m"),
-			corev1.ResourceMemory: resource.MustParse("220Mi"),
-		},
-	}
-	dcFactory := func(resources corev1.ResourceRequirements) *appsv1.DeploymentConfig {
-		return &appsv1.DeploymentConfig{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "DeploymentConfig",
-				APIVersion: "apps.openshift.io/v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "myDC",
-				Namespace: "myNS",
-			},
-			Spec: appsv1.DeploymentConfigSpec{
-				Template: &corev1.PodTemplateSpec{
-					Spec: corev1.PodSpec{
-						Containers: []corev1.Container{
-							corev1.Container{
-								Name:      "container1",
-								Resources: resources,
-							},
-						},
-					},
-				},
-			},
-		}
-	}
-
-	cases := []struct {
-		testName          string
-		existingResources corev1.ResourceRequirements
-		desiredResources  corev1.ResourceRequirements
-		expectedResult    bool
-	}{
-		{"NothingToReconcile", emptyResourceRequirements, emptyResourceRequirements, false},
-		{"NothingToReconcileWithResources", notEmptyResources, notEmptyResources, false},
-		{"AddResources", emptyResourceRequirements, notEmptyResources, true},
-		{"RemoveResources", notEmptyResources, emptyResourceRequirements, true},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.testName, func(subT *testing.T) {
-			existing := dcFactory(tc.existingResources)
-			desired := dcFactory(tc.desiredResources)
-			update := DeploymentConfigReconcileContainerResources(desired, existing, logger)
-			if update != tc.expectedResult {
-				subT.Fatalf("result failed, expected: %t, got: %t", tc.expectedResult, update)
-			}
-			if !helper.CmpResources(&existing.Spec.Template.Spec.Containers[0].Resources, &desired.Spec.Template.Spec.Containers[0].Resources) {
-				subT.Fatal(cmp.Diff(existing.Spec.Template.Spec.Containers[0].Resources, desired.Spec.Template.Spec.Containers[0].Resources, cmpopts.IgnoreUnexported(resource.Quantity{})))
-			}
-		})
-	}
-}
