--- pkg/3scale/amp/operator/route_reconciler_test.go
@@ -1,233 +1,0 @@
-package operator
-
-import (
-	"context"
-	"testing"
-
-	appsv1alpha1 "github.com/3scale/3scale-operator/pkg/apis/apps/v1alpha1"
-	routev1 "github.com/openshift/api/route/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-)
-
-func TestRouteBaseReconcilerCreate(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := routev1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	createOnlyReconciler := NewCreateOnlyRouteReconciler()
-
-	reconciler := NewRouteBaseReconciler(baseAPIManagerLogicReconciler, createOnlyReconciler)
-
-	desired := &routev1.Route{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Route",
-			APIVersion: "route.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myRoute",
-			Namespace: namespace,
-		},
-	}
-
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{
-		Name:      "myRoute",
-		Namespace: namespace,
-	}
-	existing := &routev1.Route{}
-	err = cl.Get(context.TODO(), namespacedName, existing)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestRouteBaseReconcilerUpdateOwnerRef(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	existing := &routev1.Route{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Route",
-			APIVersion: "route.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myRoute",
-			Namespace: namespace,
-		},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := routev1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existing}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	createOnlyReconciler := NewCreateOnlyRouteReconciler()
-
-	reconciler := NewRouteBaseReconciler(baseAPIManagerLogicReconciler, createOnlyReconciler)
-
-	desired := existing.DeepCopy()
-
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{
-		Name:      "myRoute",
-		Namespace: namespace,
-	}
-	reconciled := &routev1.Route{}
-	err = cl.Get(context.TODO(), namespacedName, reconciled)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(reconciled.GetOwnerReferences()) != 1 {
-		t.Fatal("reconciled does not have owner reference")
-	}
-
-	if reconciled.GetOwnerReferences()[0].Name != name {
-		t.Fatalf("reconciled owner reference is not apimanager, expected: %s, got: %s", name, reconciled.GetOwnerReferences()[0].Name)
-	}
-}
-
-type myCustomRouteReconciler struct {
-}
-
-func (r *myCustomRouteReconciler) IsUpdateNeeded(desired, existing *routev1.Route) bool {
-	existing.Spec.Path = "/newPath"
-	return true
-}
-
-func newCustomRouteReconciler() RoutesReconciler {
-	return &myCustomRouteReconciler{}
-}
-
-func TestRouteBaseReconcilerUpdateNeeded(t *testing.T) {
-	var (
-		name      = "example-apimanager"
-		namespace = "operator-unittest"
-		log       = logf.Log.WithName("operator_test")
-	)
-	apimanager := &appsv1alpha1.APIManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: appsv1alpha1.APIManagerSpec{},
-	}
-	existing := &routev1.Route{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Route",
-			APIVersion: "route.openshift.io/v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myRoute",
-			Namespace: namespace,
-		},
-	}
-	s := scheme.Scheme
-	s.AddKnownTypes(appsv1alpha1.SchemeGroupVersion, apimanager)
-	err := routev1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// existing does not need to be updated to set owner reference
-	err = controllerutil.SetControllerReference(apimanager, existing, s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existing}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewFakeClient(objs...)
-	clientAPIReader := fake.NewFakeClient(objs...)
-
-	baseReconciler := NewBaseReconciler(cl, clientAPIReader, s, log)
-	baseLogicReconciler := NewBaseLogicReconciler(baseReconciler)
-	baseAPIManagerLogicReconciler := NewBaseAPIManagerLogicReconciler(baseLogicReconciler, apimanager)
-	saReconciler := newCustomRouteReconciler()
-
-	reconciler := NewRouteBaseReconciler(baseAPIManagerLogicReconciler, saReconciler)
-
-	desired := existing.DeepCopy()
-
-	err = reconciler.Reconcile(desired)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	namespacedName := types.NamespacedName{
-		Name:      "myRoute",
-		Namespace: namespace,
-	}
-	reconciled := &routev1.Route{}
-	err = cl.Get(context.TODO(), namespacedName, reconciled)
-	// object must exist, that is all required to be tested
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if reconciled.Spec.Path != "/newPath" {
-		t.Fatalf("reconciled have reconciled data. Expected: '/newPath', got: %s", reconciled.Spec.Path)
-	}
-}
