+++ pkg/reconcilers/deploymentconfig_test.go
@@ -1,0 +1,128 @@
+package reconcilers
+
+import (
+	"testing"
+
+	"github.com/3scale/3scale-operator/pkg/helper"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+	appsv1 "github.com/openshift/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func TestDeploymentConfigReplicasReconciler(t *testing.T) {
+	dcFactory := func() *appsv1.DeploymentConfig {
+		return &appsv1.DeploymentConfig{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "DeploymentConfig",
+				APIVersion: "apps.openshift.io/v1",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "myDC",
+				Namespace: "myNS",
+			},
+			Spec: appsv1.DeploymentConfigSpec{
+				Replicas: 3,
+			},
+		}
+	}
+
+	cases := []struct {
+		testName       string
+		desired        func() *appsv1.DeploymentConfig
+		expectedResult bool
+	}{
+		{"NothingToReconcile", func() *appsv1.DeploymentConfig { return dcFactory() }, false},
+		{"ReplicasReconcile",
+			func() *appsv1.DeploymentConfig {
+				desired := dcFactory()
+				desired.Spec.Replicas = desired.Spec.Replicas + 1000
+				return desired
+			}, true,
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.testName, func(subT *testing.T) {
+			existing := dcFactory()
+			update := DeploymentConfigReplicasReconciler(tc.desired(), existing)
+			if update != tc.expectedResult {
+				subT.Fatalf("result failed, expected: %t, got: %t", tc.expectedResult, update)
+			}
+			if existing.Spec.Replicas != tc.desired().Spec.Replicas {
+				subT.Fatalf("replica reconciliation failed, existing: %d, desired: %d", existing.Spec.Replicas, tc.desired().Spec.Replicas)
+			}
+
+		})
+	}
+}
+
+func TestDeploymentConfigContainerResourcesReconciler(t *testing.T) {
+	emptyResourceRequirements := corev1.ResourceRequirements{
+		Limits:   corev1.ResourceList{},
+		Requests: corev1.ResourceList{},
+	}
+	notEmptyResources := corev1.ResourceRequirements{
+		Limits: corev1.ResourceList{
+			corev1.ResourceCPU:    resource.MustParse("100m"),
+			corev1.ResourceMemory: resource.MustParse("110Mi"),
+		},
+		Requests: corev1.ResourceList{
+			corev1.ResourceCPU:    resource.MustParse("200m"),
+			corev1.ResourceMemory: resource.MustParse("220Mi"),
+		},
+	}
+	dcFactory := func(resources corev1.ResourceRequirements) *appsv1.DeploymentConfig {
+		return &appsv1.DeploymentConfig{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "DeploymentConfig",
+				APIVersion: "apps.openshift.io/v1",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "myDC",
+				Namespace: "myNS",
+			},
+			Spec: appsv1.DeploymentConfigSpec{
+				Template: &corev1.PodTemplateSpec{
+					Spec: corev1.PodSpec{
+						Containers: []corev1.Container{
+							corev1.Container{
+								Name:      "container1",
+								Resources: resources,
+							},
+						},
+					},
+				},
+			},
+		}
+	}
+
+	cases := []struct {
+		testName          string
+		existingResources corev1.ResourceRequirements
+		desiredResources  corev1.ResourceRequirements
+		expectedResult    bool
+	}{
+		{"NothingToReconcile", emptyResourceRequirements, emptyResourceRequirements, false},
+		{"NothingToReconcileWithResources", notEmptyResources, notEmptyResources, false},
+		{"AddResources", emptyResourceRequirements, notEmptyResources, true},
+		{"RemoveResources", notEmptyResources, emptyResourceRequirements, true},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.testName, func(subT *testing.T) {
+			existing := dcFactory(tc.existingResources)
+			desired := dcFactory(tc.desiredResources)
+			update := DeploymentConfigContainerResourcesReconciler(desired, existing)
+			if update != tc.expectedResult {
+				subT.Fatalf("result failed, expected: %t, got: %t", tc.expectedResult, update)
+			}
+			if !helper.CmpResources(&existing.Spec.Template.Spec.Containers[0].Resources, &desired.Spec.Template.Spec.Containers[0].Resources) {
+				subT.Fatal(cmp.Diff(existing.Spec.Template.Spec.Containers[0].Resources, desired.Spec.Template.Spec.Containers[0].Resources, cmpopts.IgnoreUnexported(resource.Quantity{})))
+			}
+		})
+	}
+}
